Moved all typestate connection related code to a different module
