+++ aries_vcx/src/protocols/typestate_con/trait_bounds.rs
